backend/requirements.txt
langchain==0.1.14
langchain-core==0.1.52
langchain-anthropic==0.1.7
anthropic==0.26.1
httpx==0.27.0
fastapi==0.110.1
uvicorn==0.29.0
googlemaps==4.10.0
python-dotenv
